Add files via upload
Changed the ahk script to fire Encore in Chrome instead of IE.  IE was causing problems with facets displaying incorrectly, or not displaying at all.  Plus we needed to get rid of it due to Microsoft's push to get folks off of IE and onto Edge.  The switches in the code launch Chrome in incognito mode, so no cleanup of cache/cookies is necessary.